Add escape hatch and a deadline for withdrawals
The specific of these still need to be thought about more but we are
adding:

1. An escape hatch that can be called by the owner at any time. This is
a last resort call to use only in the case of an emergency. That would
in essence kill the contract and withdraw the remaining funds to a
trusted multisig.

This is to be used only if a major security issue is uncovered on the
contract or if the contract is under attack.

2. After some time (currently set to ~6 months) the remaining
non-claimed funds can be reclaimed. The party to reclaim those funds
could be the WHG, or a foundation or a charity. This is still to be decided.