Sending with 0 donation % now works
Within moments of deploying the contract, we discovered an irregular behaviour when people were donating 0% We decided to immediately stop the contract using the Escape Hatch while we looked into the irregularities.

After investigation, it appears that using the `send()` function with a value of 0 from a contract account to a multisig account will not work. The reason is that send() with a value of 0 gives a gas stipend of 0 gas which when sending to the typical multisig wallet hits this line (https://github.com/ethereum/dapp-bin/blob/master/wallet/wallet.sol#L336) and has no gas to spend. As a result, the whole transaction was throwing. This issue was fixed by not calling the send() function if a zero donation is detected. We of course had tests available (https://github.com/BitySA/whetcwithdraw/blob/086170f38eef1762726c4109a3ed713d79b87c26/tests/deploy.js#L274) for this scenario, but with a normal end-user account as the donation address instead of a multi-sig wallet. This is why the discrepancy was not discovered until we went live.